+++ apps/web/src/services/asset-inventory/collector/collector-plugins-old/modules/PluginFilter.vue
