+++ apps/web/src/services/asset-inventory/collector/modules-old/CollectionRule.vue
