+++ apps/web/src/services/asset-inventory/collector/modules-old/CollectorDetails.vue
